twitter_api/requirements.txt
alembic==1.6.0
amqp==5.0.6
billiard==3.6.4.0
biopython==1.78
celery==5.1.0b2
certifi==2020.12.5
cftime==1.4.1
chardet==4.0.0
click==8.0.0rc1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==5.0.7
et-xmlfile==1.1.0
Flask==2.0.0rc2
Flask-GraphQL==2.0.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
flask-oidc==1.4.0
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
flower==0.9.7
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
googlemaps==4.4.5
graphql-core==3.1.4
graphql-server-core==2.0.0
greenlet==1.0.0
httplib2==0.19.1
humanize==3.5.0
idna==3.1
itsdangerous==2.0.0rc2
jdcal==1.4.1
Jinja2==3.0.0rc2
joblib==1.0.1
kombu==5.1.0b1
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.0rc2
marshmallow==3.11.1
marshmallow-sqlalchemy==0.25.0
netCDF4==1.5.6
nltk==3.6.2
numpy==1.20.2
oauth2client==4.1.3
openpyxl==3.0.7
pandas==1.2.4
pbkdf2==1.3
pgeocode==0.3.0
Pillow==8.2.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
pytz==2021.1
ratelim==0.1.6
redis==3.5.3
regex==2021.4.4
requests==2.25.1
require==0.1.1
requires.io==0.2.6
rsa==4.7.2
Rx==3.2.0
six==1.15.0
SQLAlchemy==1.4.13
tornado==6.1
tqdm==4.60.0
urllib3==1.26.4
vine==5.0.0
waitress==2.0.0
wcwidth==0.2.5
Werkzeug==2.0.0rc5
xlrd==2.0.1
XlsxWriter==1.4.0
xmljson==0.2.1